DTOs & Data Annotation

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -4,6 +4,8 @@ using Microsoft.AspNetCore.Mvc;
 using ProEventos.Domain;
 using ProEventos.Application.Servicos;
 using Microsoft.AspNetCore.Http;
+using System.Collections.Generic;
+using ProEventos.Application.Dtos;
 
 namespace ProEventos.API.Controllers
 {
@@ -24,7 +26,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if(eventos == null) return NotFound("Nenhum evento encontrado");
+                if(eventos == null) return NoContent();
 
                 return Ok(eventos);
             }
@@ -42,7 +44,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if(evento == null) return NotFound("Evento não encontrado");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -60,7 +62,7 @@ namespace ProEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if(evento == null) return NotFound("Nenhum evento encontrado pelo tema");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -73,12 +75,12 @@ namespace ProEventos.API.Controllers
 
         [HttpPost]
 
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.AddEventos(model);
-                if(evento == null) return BadRequest("Erro ao tentar adicionar evento");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -91,12 +93,12 @@ namespace ProEventos.API.Controllers
 
         [HttpPut("{id}")]
 
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.UpdateEvento(id, model);
-                if(evento == null) return BadRequest("Erro ao tentar adicionar evento");
+                if(evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -113,10 +115,13 @@ namespace ProEventos.API.Controllers
         {
             try
             {
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+                if(evento == null) return NoContent();
+
                 if(await _eventoService.DeleteEvento(id))
                     return Ok("Deletado");
                 else
-                    return BadRequest("Evento não deletado");
+                    throw new Exception("Erro ao tentar deletar evento.");
             }
             catch (Exception ex)
             {
